requirements.txt
gunicorn
flask
tensorflow==2.13.0
pillow
numpy
